==> codi/2.PREPARACIO.Rmd <==
---
title: 'PRECAV: Caracterización del riesgo cardiovascular a una edad prematura
  en Cataluña (PREPARACIÓ)'
author: "Jordi Real"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: paged
  pdf_document: default
  word_document: default
params:
  evaluar: TRUE 
  fitxer_entrada: "BD_PRECAV_test5.rds" 
  filtre: "c_inclusio_grup_MACE>=0" 
  mostra: yes # no
  
  
---


# Càrrega de funcions  

```{r setup, include=T}
library(knitr)
knitr::opts_chunk$set(
	eval = params$evaluar,
	echo = T,
	error = FALSE,
	warning = FALSE,
	include = T
)



library(dplyr)
library(here)


```

# 1. DIRECTORI DE TREBALL              

>> Inicialització de parametres  

```{r inicialitzacio, include=T}

if (params$mostra) dir_preparades<-"dades/preparades/test" else dir_preparades<-"dades/preparades"


fitxer_input<-here::here(dir_preparades,"BD_PRECAV.rds")
# fitxer_input<-here::here(dir_preparades,params$fitxer_entrada)


dades_flow_input<-here::here(dir_preparades,"dades_flow.rds")
# dades_flow_input<-here::here("dades","dades_flow.rds")


link_source<-paste0("https://github.com/jrealgatius/Stat_codis/blob/master/funcions_propies.R","?raw=T")
devtools::source_url(link_source)


conductor_variables<-here::here("variables_precav.xls")


```


# 1. Lectura de fitxer

```{r lectura}

dades<-readRDS(fitxer_input) %>% as_tibble()

dades_flow<-readRDS(dades_flow_input) 


```


# 1.2. Generació de conjunt de dades imputades

```{r imputacio, eval=F}

# Fer IMPUTACIO I GUARDAR LONG + DATASET INICIAL  ----------------------

# Mice metode
library(mice)
# 1. Selecciono data_set de variables que vull utilitzar per la imputació   -------
variables_a_imputar<-extreure.variables("imp",taulavariables = conductor_variables)

dades_imputation<-dades %>% select("idp",variables_a_imputar)
dades_no_imputades<-dades %>% select(-variables_a_imputar,"idp")

# Aplico model d'imputació  (Genero les dades imputades)  ----------
num_imputacions<-1 # Normalment >=10
num_iteracions<-5 # Normalment 5
# Generar imputacions
start_time <- Sys.time()
# mice(data=dades_imputation,m=num_imputacions,maxit=5,meth='pmm',seed=123,remove.collinear=T)
tempData <- mice(data=dades_imputation,m=num_imputacions,maxit=num_iteracions,meth='cart',seed=123,remove.collinear=T)
# ,remove.collinear=F)
end_time <- Sys.time()
end_time-start_time

# Generar llista de dades imputades en llista i en data_frame
data_list <- 1:num_imputacions %>% map(~mice::complete(tempData, .x)) 
data_long<-mice::complete(tempData, action = "long", include = F)
data_long_total<-mice::complete(tempData, action = "long", include = T)

# Juntar data_long_total les dades no imputades
dades<-data_long_total %>% left_join(dades_no_imputades,by="idp")
# Convertir objecte MICE
miceData<-as.mids(data_long_total)

```



# 2. Calculs i Recodificacions  ----------------


```{r calcular_tempsDM}
# 

library(lubridate)

# Incloure dins de DM2 (DG.E11) Altres DM (DG.DM_O)
dades<-dades %>% mutate(DG.E11=ifelse(!is.na(DG.DM_O) & is.na(DG.E11),DG.DM_O,DG.E11))

# Calcul de temps DM2
dades<-dades %>% 
  mutate(Duracion_DM2 = as.period(interval(start = ymd(DG.E11), end = ymd(dtindex))) %>% time_length("year")) 

# Calcul de temps DM1
dades<-dades %>% 
  mutate(Duracion_DM1 = as.period(interval(start = ymd(DG.E10), end = ymd(dtindex))) %>% time_length("year")) 

```


```{r calculs1}
library(lubridate)


##  Calcular edat 
dades<-dades %>% 
  mutate(edat = lubridate::year(as.period(interval(start = ymd(dnaix), end = dtindex))))

# Farmacs 
# NA --> 0 (Else=1) (No hi ha 0'0)
dades<-dades %>% mutate_at(vars(starts_with("FP.")),
                           ~if_else(is.na(.) | .==0,0,1)) 

dades<-dades %>% mutate_at(vars(starts_with("FP1_3a.")), 
                           ~if_else(is.na(.) | .== 0,0,1)) 

# NA --> 0 (ALTRI ES UNA COPIA)
dades<-dades %>% mutate_at(vars(starts_with("FD.")), 
                           ~if_else(is.na(.) | .==0,0,1)) 

# NA --> 0 (ALTRI ES UNA COPIA)
dades<-dades %>% mutate_at(vars(starts_with("FD1_3a.")), 
                           ~if_else(is.na(.) | .==0,0,1)) 

# Diagnostics
# NA --> 0 (Else=1) (No hi ha 0)
dades<-dades %>% mutate_at(vars(starts_with("DG.")), 
                           ~if_else(is.na(.) | .==0,0,1)) 

# Diagnostics (ANTECEDENTS)
# NA --> 0 (Else=1) (No hi ha 0)
dades<-dades %>% mutate_at(vars(starts_with("ANT1_3.")), 
                           ~if_else(is.na(.) | .==0,0,1)) 

# Events
dades<-dades %>% mutate_at(vars(starts_with("EV.")), 
                           ~if_else(is.na(.) | .==0,0,1) ) 

# Càlcul de número d'events Terciaris
dades<-dades %>% mutate(
  EV.TER_SUM = rowSums(select(., starts_with("EV.TER."))))

# Recode EV.TER_SUM
dades<-dades %>% mutate(
  EV.TER_multi = case_when(
    EV.TER_SUM==1~"1 terr",
    EV.TER_SUM>1~">=2 terr")
  )

# Càlcul de número d'events tipus
dades<-dades %>% mutate(
  EV.TIP_SUM = rowSums(select(., starts_with("EV.TIP.")))
  )

# Recode EV.TIP_SUM
dades<-dades %>% mutate(
  EV.TIP_multi = case_when(
  EV.TIP_SUM==1~"1 ECV",
  EV.TIP_SUM>1~">=2 ECVs")
  )


# visites 
dades<-dades %>% mutate_at(vars(starts_with("visites_")), 
                           ~ifelse(is.na(.) | .==0,0,.) )

# Incloure dins de DM2 (DG.E11) Altres DM (DG.DM_O)
dades<-dades %>% mutate(DG.E11=ifelse(DG.DM_O==1,1,DG.E11),
                 ANT1_3.E11=ifelse(ANT1_3.DM_O==1,1,ANT1_3.E11))


```

```{r calculs2}
# Recodes
dades<-dades %>% 
  mutate(edat_pre=ifelse(sexe=="H" & edat<55,"Young man(<55)",NA)) %>% 
  mutate(edat_pre=ifelse(sexe=="H" & edat>=55,"Old man(>=55)",edat_pre)) %>% 
  mutate(edat_pre=ifelse(sexe=="D" & edat<65,"Young woman(<65)",edat_pre)) %>% 
  mutate(edat_pre=ifelse(sexe=="D" & edat>=65,"Old woman(>=65)",edat_pre)) 

# Recodes
dades<-dades %>% 
  mutate(edat_grup=ifelse((sexe=="H" & edat<55) | (sexe=="D" & edat<65) ,"Young:M<55/F<65",NA)) %>% 
  mutate(edat_grup=ifelse((sexe=="H" & edat>=55) | (sexe=="D" & edat>=65),"Old:M>=55/F>=65",edat_grup))  


# Recodes
dades<-dades %>% 
  mutate(edat_grup2=ifelse(edat<35,1,NA)) %>% 
  mutate(edat_grup2=ifelse(sexe=="H" & edat>=35 & edat<55,2,edat_grup2)) %>% 
  mutate(edat_grup2=ifelse(sexe=="D" & edat>=35 & edat<60,2,edat_grup2)) %>% 

  mutate(edat_grup2=ifelse(sexe=="H" & edat>=55 & edat<65,3,edat_grup2)) %>% 
  mutate(edat_grup2=ifelse(sexe=="D" & edat>=60 & edat<65,3,edat_grup2)) %>% 

  mutate(edat_grup2=ifelse(sexe=="H" & edat>=65 & edat<75,4,edat_grup2)) %>% 
  mutate(edat_grup2=ifelse(sexe=="D" & edat>=65 & edat<75,4,edat_grup2)) %>% 

  mutate(edat_grup2=ifelse(edat>=75 ,5,edat_grup2))
```


```{r calculs2.2}

# Recodes
dades<-dades %>% 
  mutate(any_index=factor(lubridate::year(dtindex)))

# REGICOR  ------------------

# Selecciono variables REGICOR i formatejo 
# Ojo que s'ha de recalcular la edat 1-3 anys previs 
# S'ha eliminat ANT1_3.E10 (Ja no existeix)

dades_regicor<-selectorvariables("regicor",taulavariables=conductor_variables,dt=dades) %>% 
  mutate(age=edat,
         diabetes=if_else(ANT1_3.E11==1 | ANT1_3.E10==1,1,0),
         smoker=if_else(tab.valor==1,1,0,missing = 0), 
         coltot=COLTOT.valor13a,
         colhdl=COLHDL.valor13a,
         sbp=PAS.valor13a,
         dbp=PAD.valor13a) 
# Aplico formula i ho fuciono a dades
temp<-dades_regicor
regicor_df<-regicor(temp$age,temp$sexe,temp$smoker,temp$diabetes,temp$coltot,temp$colhdl,temp$sbp,temp$dbp) 

regicor_df<-tibble(regicor_df) %>% rename(regicor=regicor_df)

dades<-dades %>% bind_cols(regicor_df)

rm(temp,dades_regicor,regicor_df)

```

```{r calcul_score2}

link_source<-paste0("https://github.com/jrealgatius/Stat_codis/blob/master/Scores.R","?raw=T")
devtools::source_url(link_source)


dades<-dades %>% mutate(
  score2=
    score2(sex=sexe,age=edat,
           smoke =tab.valor,
           sbp=PAS.valor13a,
           dm=as.numeric(ANT1_3.E11==1 | ANT1_3.E10==1),
           coltot=COLTOT.valor13a,
           colhdl=COLHDL.valor13a,
           risk_region = "Low")) 


```


```{r calculs3}
# Variables Noves Emilio Ortega --------------
# HDL BAJO:  hdlbajo13a_cat2 (1-3a=ultims 3 anys)

dades<-dades %>% 
  mutate(hdlbajo13a_cat2=case_when(
    sexe=="H" & COLHDL.valor13a < 40 ~ "Si",
    sexe=="D" & COLHDL.valor13a < 50 ~ "Si",
    sexe=="H" & COLHDL.valor13a >=40 ~ "No",
    sexe=="D" & COLHDL.valor13a >=50 ~ "No"))

# Dislipemia aterogénica disl13a_cat2  (Any Missings = missings)
dades<-dades %>% 
  mutate(disl13a_cat2=case_when(
    TG.valor13a>200 & hdlbajo13a_cat2=="Si"~ "Si",
    TG.valor13a>200 & hdlbajo13a_cat2=="No"~ "No",
    TG.valor13a<=200 & hdlbajo13a_cat2=="Si"~ "No",
    TG.valor13a<=200 & hdlbajo13a_cat2=="No"~ "No")) 
                 
# Colesterol remanente: col_rema13a
dades<-dades %>% mutate(col_rema13a=COLTOT.valor13a-COLLDL.valor13a-COLHDL.valor13a) 

# Colesterol noHDL: col_noHDL13a 
dades<-dades %>% mutate(col_noHDL13a=COLTOT.valor13a--COLHDL.valor13a) 

# regicor_cat2 (DM1 no te regicor)
dades<-dades %>% mutate(regicor_cat2=case_when(
  regicor<5~1,
  regicor>=5 & regicor<10~2,
  regicor>=10 & regicor<15~3,
  regicor>=15 ~4)) 

# regicor_cat (DM1 no te regicor)
dades<-dades %>% mutate(regicor_cat=case_when(
  regicor<5~1,
  regicor>=5 & regicor<10~2,
  regicor>=10 ~3)) 


# regicor_mis 
dades<-dades %>% mutate(regicor_mis=if_else(is.na(regicor),"Missing","Value") %>% as.factor())

# Recodes to missings llista de vars 
llista_vars<-c("COLHDL.valor13a","COLTOT.valor13a","PAD.valor13a","PAS.valor13a","tab.valor")
temp_dt<-dades %>% select(llista_vars) %>% mutate_all(~if_else(is.na(.) | .==0,"Missing","Value")) 
colnames(temp_dt)<-paste0(llista_vars,".mis")
dades<-dades %>% cbind(temp_dt) %>% as_tibble()

# FHF_cat (fenotipo hipercolesterolemia familiar) -----
dades<-dades %>% 
  mutate(FHF_cat=case_when(
    COLLDL.valor13a>=230 & edat<30~1,
    COLLDL.valor13a>=238 & edat>=30 & edat<39~1,
    COLLDL.valor13a>=260 & edat>=40 & edat<49~1,
    COLLDL.valor13a>=265 & edat>=49~1)) %>% 
  mutate(FHF_cat=case_when(
    !is.na(COLLDL.valor13a) & is.na(FHF_cat)~0,
    !is.na(COLLDL.valor13a) & FHF_cat==1~1))

# DG.HTA_O + DG.COMP_HTA ----------------
dades<-dades %>% mutate(DG.HTA=if_else(DG.HTA_O==1 | DG.COMP_HTA==1,1,0))

# LDL ALTO: ldlalto13a_cat2 (LDL>190 sin tratamiento hipilipemiante)
dades<- dades %>% 
  mutate(ldlalto13a_cat2=case_when(
    COLLDL.valor13a>190 & FD1_3a.HIPOLIP==0 ~ "LDL>190 Sin Tx",
    COLLDL.valor13a>190 & FD1_3a.HIPOLIP==1 ~ "LDL>190 con Tx",
    COLLDL.valor13a<=190 ~ "LDL<190"))

# LDL BAJO: ldlbajo13a_cat2 (LDL>160 con tratamiento hipilipemiante)
dades<- dades %>% 
  mutate(ldlbajo13a_cat2=case_when(
    COLLDL.valor13a>160 & FD1_3a.HIPOLIP==1 ~ "LDL>160 + Tx",
    COLLDL.valor13a>160 & FD1_3a.HIPOLIP==0 ~ "LDL>160 sin Tx",
    COLLDL.valor13a<=160 ~ "LDL<160"))

# Variable seguimento -----------------
dades<-dades %>% mutate (tmps_seguiment=dtindex-lubridate::dmy("01012007"),
                         anys_seguiment=as.numeric(tmps_seguiment)/365.25)


# Edat/homes i dones
dades<-dades %>% mutate(edatH=ifelse(sexe=="H",edat,NA))
dades<-dades %>% mutate(edatD=ifelse(sexe=="D",edat,NA))

# Dislipemia ((TG>150 | COLTOTAL>240 ) & HIPOLIPEMIANTES (FD.HIPOLIP PRESCRITOS O DISPENSADOS))
# TG.valor COLTOT.valor
# FD.HIPOLIP FP.HIPOLIP 

dades<-dades %>% mutate(DG.DSL=if_else(FD.HIPOLIP | FP.HIPOLIP | TG.valor>150 | COLTOT.valor>240,1,0)) %>% 
  mutate(DG.DSL=if_else(is.na(DG.DSL),0,DG.DSL)) 


# Generar variable event primari grup a risk: (seleccionar grups a risk només amb event primari)
dades<- dades %>% group_by(caseid) %>% mutate(EV.PRM_grup=ifelse(sum(EV.ECV_PR==1),1,0)) %>% ungroup() 




```



```{r}
# Recodes automatics
dades<-recodificar(dades,taulavariables = conductor_variables,criteris = "recodes")
dades<-recodificar(dades,taulavariables = conductor_variables,criteris = "recodes2")


# Generar filtre posterior (Grups , Event i seus controls: només si es MACE (s'exclou: nomes Insuficiecia cardiaca / Artper) 
dades<-dades %>% group_by(caseid) %>% mutate(c_inclusio_grup_MACE=ifelse(sum(EV.ECV_AMP==1),1,0)) %>% ungroup()

# Visites previes: 1.Si/0.No
dades<-dades %>% mutate(visites_cat=ifelse(visites_TOT>=1,1,0))


```


# 3. Criteris d'exclusió post / dades per flow-chart

** Ojo que només s'excluen FA / i per edat i grups corresponents sense control
** S'exclou tot el grups 

* Casos sense control
* Edad inferior  35 o superior o igual a75 anys   
* Amb FA en data index
* Descartat filtre per DM1 

```{r filtratje}
# Dades flow-chart
# Aplicar filtre inicial de grup
dades<-dades %>% filter(eval(parse(text=params$filtre)))

# Generar criteris d'exclusió grupals (Si ho cumpleix condició eliminar tot el grup)

# Exclusió per edat
dades<-dades %>% mutate(c_exclusio_edat=ifelse(edat<35 | edat>=75,1,0)) %>% 
  group_by(caseid) %>% mutate(c_exclusio_edat=sum(c_exclusio_edat)) %>% ungroup() 
# Exclusió per FA
dades<-dades %>% mutate(c_exclusio_FA=ifelse(DG.FA==1,1,0)) %>% group_by(caseid) %>% mutate(c_exclusio_FA=sum(c_exclusio_FA)) %>% ungroup()


# Genero flow-chart sequencial 
flow_chart<-criteris_exclusio_diagrama(dt=dades,taulavariables = conductor_variables,
                                       criteris = "c_exclusio1",etiquetes = "exc_lab",
                                       grups = "event",ordre="exc_ordre",sequencial=T,
                                       pob_lab=c("Pob aparellada nascuda: 1935-1981","Pob final"))

# Aplicar criteris d'exclusió
dades<-dades %>% criteris_exclusio(conductor_variables,criteris = "c_exclusio1")

# Verificació de filtres i salvar base d'events
events_dt <-dades %>% filter(event==1) 
events_dt %>% skimr::skim(edat)



```



# 4.1 Generar Flow-chart 1

```{r flow_chart, include=F}
# Generar Flow_chart 


# Configuro parametres del flowchart
pob_lab<-"Población inicial (8%)"
pob_lab1<-c("Potenciales controles","Controles")
pob_lab2=c("Potenciales casos","Casos")

# N inicial
Ninicial<-dades_flow %>% filter(etiqueta=="Inicial") %>% pull(N) %>% sum()

# N inicial final per grup
NControls_inicial_final<-dades_flow %>% filter(grup=="Control" & (etiqueta=="Inicial" | etiqueta=="Post_Match")) %>% pull(N) 
NCas_inicial_final<-dades_flow %>% filter(grup=="Cas" & (etiqueta=="Inicial" | etiqueta=="Post_Match")) %>% pull(N) 

# Exclusions controls
Nexcl_controls<-dades_flow %>% filter(grup=="Control" & tipo=="exc") %>% pull(N) 
Labexcl_controls<-dades_flow %>% filter(grup=="Control" & tipo=="exc") %>% pull(etiqueta) 

# Exclusions cas
Nexcl_cas<-dades_flow %>% filter(grup=="Cas" & tipo=="exc") %>% pull(N) 
Labexcl_cas<-dades_flow %>% filter(grup=="Cas" & tipo=="exc") %>% pull(etiqueta) 


flow_chart_antic<-diagramaFlowchart2G(pob_lab=pob_lab,
                                pob=Ninicial,
                                pob_lab1=pob_lab1,
                                pob_lab2=pob_lab2,
                                
                                pob1=NControls_inicial_final, # Controls inicials i finals  
                                pob2=NCas_inicial_final, # Casos inicials i finals 
                                
                                exc1=Nexcl_controls,
                                exc_lab1=Labexcl_controls,
                                exc2=Nexcl_cas,
                                exc_lab2=Labexcl_cas)



###  Genero dades flowchat 2 excloent els events Insf cardiaca  ##

```

# 4.2. Generar Flow-chart 2: Excloent grups amb Insfuciencia cardiaca

```{r flow_chart2,include=F}
# Generar Flow_chart 


# Configuro parametres del flowchart
pob_lab<-"Población inicial (8%)"
pob_lab1<-c("Potenciales controles","Controles")
pob_lab2=c("Potenciales casos","Casos")

# N inicial
Ninicial<-dades_flow %>% filter(etiqueta=="Inicial") %>% pull(N) %>% sum()

# N inicial final per grup
NControls_inicial_final<-dades_flow %>% filter(grup=="Control" & (etiqueta=="Inicial" | etiqueta=="Post_Match")) %>% pull(N) 
NCas_inicial_final<-dades_flow %>% filter(grup=="Cas" & (etiqueta=="Inicial" | etiqueta=="Post_Match")) %>% pull(N) 

# Exclusions controls
Nexcl_controls<-dades_flow %>% filter(grup=="Control" & tipo=="exc") %>% pull(N) 
Labexcl_controls<-dades_flow %>% filter(grup=="Control" & tipo=="exc") %>% pull(etiqueta) 

# Exclusions cas
Nexcl_cas<-dades_flow %>% filter(grup=="Cas" & tipo=="exc") %>% pull(N) 
Labexcl_cas<-dades_flow %>% filter(grup=="Cas" & tipo=="exc") %>% pull(etiqueta) 


flow_chart_antic2<-diagramaFlowchart2G(pob_lab=pob_lab,
                                pob=Ninicial,
                                pob_lab1=pob_lab1,
                                pob_lab2=pob_lab2,
                                
                                pob1=NControls_inicial_final, # Controls inicials i finals  
                                pob2=NCas_inicial_final, # Casos inicials i finals 
                                
                                exc1=Nexcl_controls,
                                exc_lab1=Labexcl_controls,
                                exc2=Nexcl_cas,
                                exc_lab2=Labexcl_cas)

flow_chart_antic2


###  Genero dades flowchat 2 excloent els events Insf cardiaca  ##

```

# 4. Etiquetació / factorització / relevels

Factoritzar NO.Yes llista de variables "factor" situades a la taulavariables camp=factor


```{r etiquetacio}

vars_factor<-read_conductor(conductor_variables) %>% filter(factor.YESNO>0) %>% pull(camp)

dades<-dades %>% mutate_at(vars_factor,~factor(.,levels=c(0,1), labels=c("No","Yes"))) 
  
# Value labels
dades<-etiquetar_valors(dades,variables_factors = conductor_variables,fulla="value_label")

# Factoritzar variables character (excepte c("idp","idup", "codindex"))
vars_char<-dades %>% select_if(is.character) %>% select(-c("idp","idup", "codindex")) %>% names()
dades<-dades %>% mutate_at(vars_char,as.factor)

# Etiquetar 
dades<-etiquetar(d=dades,taulavariables = conductor_variables)

# Relevel FG de mes gran a més petit
dades<-dades %>% mutate(CKDEPI.valor13a.cat4=factor(CKDEPI.valor13a.cat4,levels = rev(levels(CKDEPI.valor13a.cat4))))

# Relevel MEDEA U1 Referencia 
dades<-dades %>% mutate(qmedea=factor(qmedea,levels = rev(levels(qmedea))))
dades<-dades %>% mutate(qmedea=relevel(qmedea,"U1"))


```


# 6. Analisis preliminar / Objectiu secundari 

>- Descripció dels events per edad i sexe i en funció de si son prematurs o no

```{r descriptiva}
library(compareGroups)

dadesevents<-dades %>% filter(event==1 | event=="Caso")

# Per genere
formula<-formula.text(x="events",y="sexe",taulavariables = conductor_variables) 
T1.EVENTS<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)

formula<-formula_compare(x="events_ter",y="sexe",taulavariables = conductor_variables) 
T1.2.EVENTS<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)

formula<-formula_compare(x="events_imp",y="sexe",taulavariables = conductor_variables) 
T1.3.EVENTS<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)

# Prematur (Si/No)
formula<-formula_compare(x="events",y="edat_grup",taulavariables = conductor_variables) 
T1.EVENTS.2<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = F)

formula<-formula_compare(x="events_ter",y="edat_grup",taulavariables = conductor_variables) 
T1.2.EVENTS.2<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)

formula<-formula_compare(x="events_imp",y="edat_grup",taulavariables = conductor_variables) 
T1.3.EVENTS.2<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)

# Prematur (Si/No) + Sexe
formula<-formula_compare(x="events",y="edat_pre",taulavariables = conductor_variables) 
T1.EVENTS.3<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = F)

formula<-formula_compare(x="events_ter",y="edat_pre",taulavariables = conductor_variables) 
T1.2.EVENTS.3<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = F)

formula<-formula_compare(x="events_imp",y="edat_pre",taulavariables = conductor_variables) 
T1.3.EVENTS.3<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = F)

# Per tres grups d'edat
formula<-formula_compare(x="events",y="edat.cat5",taulavariables = conductor_variables) 
T1.7.EVENTS<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)

formula<-formula_compare(x="events_ter",y="edat.cat5",taulavariables = conductor_variables) 
T1.8.EVENTS<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)


formula<-formula_compare(x="events_imp",y="edat.cat5",taulavariables = conductor_variables) 
T1.9.EVENTS<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)
```


```{r descriptiva2}

# Descriptiu de farmacs i Baseline
formula<-formula_compare(x="farmacs",y="",taulavariables = conductor_variables) 
T1.FARMACS<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)

formula<-formula_compare(x="baseline",y="",taulavariables = conductor_variables) 
T1.BASELINE<-descrTable(formula,data=dadesevents,show.p.overall = F,hide.no = "No",show.all = T)


```

>-  Comparativa amb controls

```{r comparativa1}
# Tecnica per no arrossegar tota la base de dades
formula<-formula_compare(x="baseline",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("baseline",conductor_variables),"event","edat_grup")
T2.BASELINE<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.all = F,show.n = T)

dades_temp<-dades %>% select(extreure.variables("baseline",conductor_variables),"event","edat_grup") %>% filter(edat_grup=="Young:M<55/F<65")
T2.BASELINE.joves<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.all = F,show.n = T)
dades_temp<-dades %>% select(extreure.variables("baseline",conductor_variables),"event","edat_grup") %>% filter(edat_grup=="Old:M>=55/F>=65")
T2.BASELINE.grans<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.all = F,show.n = T)

formula<-formula_compare(x="events",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("events",conductor_variables),"event","edat_grup")
T2.EVENTS<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No")

formula<-formula_compare(x="events",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("events",conductor_variables),"event","edat_grup") %>% filter(edat_grup=="Young:M<55/F<65")
T2.EVENTS.joves<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No")

formula<-formula_compare(x="events",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("events",conductor_variables),"event","edat_grup") %>% filter(edat_grup=="Old:M>=55/F>=65")
T2.EVENTS.grans<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No")


formula<-formula_compare(x="farmacs_previs",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("farmacs_previs",conductor_variables),"event","edat_grup")
T2.FARMACS<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No")

formula<-formula_compare(x="farmacs",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("farmacs",conductor_variables),"event","edat_grup")
T2.FARMACS_ACTUALS<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No")

dades_temp<-dades %>% select(extreure.variables("farmacs",conductor_variables),"event","edat_grup") %>% filter(edat_grup=="Young:M<55/F<65")
T2.FARMACS_ACTUALS.joves<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No") 
dades_temp<-dades %>% select(extreure.variables("farmacs",conductor_variables),"event","edat_grup") %>% filter(edat_grup=="Old:M>=55/F>=65")
T2.FARMACS_ACTUALS.grans<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No") 

formula<-formula_compare(x="antecedents_previs",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("antecedents_previs",conductor_variables),"event","edat_grup")
T2.ANTECEDENTS<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T)

formula<-formula_compare(x="valors_previs",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("valors_previs",conductor_variables),"event","edat_grup")
T2.VALORS<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T)

formula<-formula_compare(x="regicor",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("regicor",conductor_variables),"event","edat_grup")
T2.regicor<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = T)

# Per grup d'edad i genere
formula<-formula_compare(x="regicor",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("regicor",conductor_variables),"event","edat_grup")
T2.regicor<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = T)

```


```{r comparativa2}
# Grans/Joves stratificat per sexe  

# Carrego funcio strataTable de Comparegroups
source("strataTable.R")

# En formula elimino variables que no tenen efectius en algun dels grups
formula_text<-formula.text(x="baseline",y="event",taulavariables = conductor_variables,eliminar = c("edat_grup","edat_pre","DG.PIEDIAB","DG.HIPO","edat.cat5"))

dades_temp<-dades %>% select(extreure.variables("baseline",conductor_variables),"event","edat_grup") 

# Estratificat per grups
taula<-descrTable(formula_text,data=dades_temp,show.p.overall = F,show.all = F,show.n = T,simplify = F)
T2.BASELINE.sex.edat<-strataTable(taula,"edat_pre")

# Farmacs 
formula<-formula.text(x="farmacs_previs",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% select(extreure.variables("farmacs_previs",conductor_variables),"event","edat_grup","edat_pre")

taula<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No")
T2.FARMACS_ACTUALS.sex.edat<-strataTable(taula,"edat_pre")

```

>- Comparativa amb controls per grups de edad (edat_grup) 

```{r comparativa_edat}

formula<-formula_compare(x="regicor",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% filter(edat_grup=="Young:M<55/F<65") %>% select(extreure.variables("regicor",conductor_variables),"event")
T2.regicor_g1<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = T)

formula<-formula_compare(x="regicor",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% filter(edat_grup=="Old:M>=55/F>=65") %>% select(extreure.variables("regicor",conductor_variables),"event")
T2.regicor_g2<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = T)

```

## Taules per objectiu primari

>> Descriptiva i significació entre casos i controls en població jove

```{r regicor}
# Descriptiu primari ("Young:M<55/F<65")

formula<-formula_compare(x="taula1",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% filter(edat_grup=="Young:M<55/F<65") 

dades_temp<-dades_temp %>% select(extreure.variables("taula1",conductor_variables),"event","sexe")
Taula1<-descrTable(event~.-sexe,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F)
Taula1.2<-strataTable(Taula1,"sexe")

# OR's i sig primaris global

# Definició de valor del outcome
cat_event<-"Caso"

# Factoritzo novament (Sino peta)
dades<-dades %>% mutate(regicor_mis=factor(regicor_mis))


dades_temp<-dades %>% filter(edat_grup=="Young:M<55/F<65") 


M1.reg.alon_conti.ls<-extreure_model_logistic(x="regicor_alone_continu",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M1.reg.alon.ls<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M1.reg.mis.ls<-extreure_model_logistic(x="regicor_mis",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")

# Descriptiu primari per sexe 
# Homes Joves
dades_temp<-dades %>% filter(sexe=="H" & edat_grup=="Young:M<55/F<65") 
M1.reg.alon_conti.ls.H<-extreure_model_logistic(x="regicor_alone_continu",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M1.reg.alon.ls.H<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M1.reg.mis.ls.H<-extreure_model_logistic(x="regicor_mis",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")

# Dones Joves
dades_temp<-dades %>% filter(sexe=="D" & edat_grup=="Young:M<55/F<65") 
M1.reg.alon_conti.ls.D<-extreure_model_logistic(x="regicor_alone_continu",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M1.reg.alon.ls.D<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M1.reg.mis.ls.D<-extreure_model_logistic(x="regicor_mis",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")

```




>- Evaluar Regicor (Taules per objectiu primari)

>> Métode: Models Logistics condicionals de: Regicor / variables incloses en Regicor / Variables regicor + farmacs 


```{r regicor_plus}

M2.reg.vars.ls<-extreure_model_logistic(x="regicor_vars2",y="event",taulavariables=conductor_variables,dades=dades,valor_outcome=cat_event,conditional = T,strata = "caseid")
# Numero de efectius amb variables regicor
n_regicor_vars<-na.omit(selectorvariables("regicor_vars2",taulavariables = conductor_variables,dades)) %>% count()

M3.reg.plus.ls<-extreure_model_logistic(x="regicor_plus",y="event",taulavariables=conductor_variables,dades=dades,valor_outcome=cat_event,conditional = T,strata = "caseid")
# Numero de efectius amb variables regicor_plus
n_regicor_plus<-na.omit(selectorvariables("regicor_plus",taulavariables = conductor_variables,dades)) %>% count()



```


>- Evaluar Regicor GLOBAL / Subgrups  

```{r regicor_global, include=F, eval=F}
library(plotROC)
ROC.reg.alone.plot<-ggplot(dades, aes(d = event, m = regicor,color=edat_grup)) + 
  plotROC::geom_roc(n.cuts = 0,increasing=F) + 
  plotROC::style_roc()

ROC.reg.alone.plot <- ROC.reg.alone.plot + style_roc(theme = theme_grey) +
  theme(axis.text = element_text(colour = "blue")) +
  ggtitle("ROC Curve") +
  annotate("text", x = .75, y = .25,
             label = paste("AUC =", round(plotROC::calc_auc(ROC.reg.alone.plot)["AUC"], 2))) 

```

>- Evaluar Regicor Grup d'edat Joves (G1)

```{r regicor_joves, eval=T, include=F}
dades_temp<-dades %>% filter(edat_grup=="Young:M<55/F<65")
M1.reg.alon.G1.ls<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M2.reg.vars.G1.ls<-extreure_model_logistic(x="regicor_vars2",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M3.reg.plus.G1.ls<-extreure_model_logistic(x="regicor_plus",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")

```

>- Evaluar Regicor Grup d'edat Grans (G2)

```{r regicor_g2, include=F, eval=T}

dades_temp<-dades %>% filter(edat_grup=="Old:M>=55/F>=65")
M1.reg.alon.G2.ls<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M2.reg.vars.G2.ls<-extreure_model_logistic(x="regicor_vars2",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M3.reg.plus.G2.ls<-extreure_model_logistic(x="regicor_plus",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")


```

>- Evaluar Regicor Grup dones Joves (G3)

```{r regicor_g3, eval=T, include=F}
dades_temp<-dades %>% filter(edat_grup=="Young:M<55/F<65" & sexe=="D")

M1.reg.alon.G3.ls<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M2.reg.vars.G3.ls<-extreure_model_logistic(x="regicor_vars2",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M3.reg.plus.G3.ls<-extreure_model_logistic(x="regicor_plus",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")

```

>- Evaluar Regicor Grup dones Granss (G4)

```{r regicor_g4, include=F, eval=T}

dades_temp<-dades %>% filter(edat_grup=="Old:M>=55/F>=65" & sexe=="D")
M1.reg.alon.G4.ls<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M2.reg.vars.G4.ls<-extreure_model_logistic(x="regicor_vars2",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M3.reg.plus.G4.ls<-extreure_model_logistic(x="regicor_plus",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")


```

>- Evaluar Regicor Grup homes Joves (G5)

```{r regicor_dones_joves, eval=T, include=F}
dades_temp<-dades %>% filter(edat_grup=="Young:M<55/F<65" & sexe=="H")

M1.reg.alon.G5.ls<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M2.reg.vars.G5.ls<-extreure_model_logistic(x="regicor_vars2",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M3.reg.plus.G5.ls<-extreure_model_logistic(x="regicor_plus",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")

```

>- Evaluar Regicor Grup homes Granss (G6)

```{r regicor_g6, include=F, eval=T}

dades_temp<-dades %>% filter(edat_grup=="Old:M>=55/F>=65" & sexe=="H")
M1.reg.alon.G6.ls<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M2.reg.vars.G6.ls<-extreure_model_logistic(x="regicor_vars2",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")
M3.reg.plus.G6.ls<-extreure_model_logistic(x="regicor_plus",y="event",taulavariables=conductor_variables,dades=dades_temp,valor_outcome=cat_event,conditional = T,strata = "caseid")


```


Conocer el porcentaje de personas que desarrollan el primer evento cardiovascular prematuro (< 55/65 años) de los que no se dispone de información clínica previa necesaria para para evaluar su riesgo cardiovascular mediante la ecuación REGICOR en los 5 años previos al mismo

```{r secundari2 }
# Seleccionar població amb Esdeveniment població prematura i fer una descriptiva de les variables incloses en el Regicor

dades_temp<-dades %>% filter(event=="Caso" & edat_grup=="Young:M<55/F<65") %>%
  select(extreure.variables("regicor_vars_mis",conductor_variables),"event","edat_pre","sexe")
formula<-formula_compare(x="regicor_vars_mis",y="sexe",taulavariables = conductor_variables) 
taula_REGICOR_events_joves<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F,show.all = T)

taula_REGICOR_events_joves


dades_temp<-dades %>% filter(event=="Control" & edat_grup=="Young:M<55/F<65") %>%
  select(extreure.variables("regicor_vars_mis",conductor_variables),"event","edat_pre","sexe")
formula<-formula_compare(x="regicor_vars_mis",y="sexe",taulavariables = conductor_variables) 
taula_REGICOR_controls_joves<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F,show.all = T)

taula_REGICOR_controls_joves


dades_temp<-dades %>% filter(event=="Caso") %>%
  select(extreure.variables("regicor_vars_mis",conductor_variables),"event","edat_pre","sexe")
formula<-formula_compare(x="regicor_vars_mis",y="edat_pre",taulavariables = conductor_variables) 
taula_REGICOR_events_grups<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F,show.all = T)

taula_REGICOR_events_grups

dades_temp<-dades %>% filter(event=="Control") %>%
  select(extreure.variables("regicor_vars_mis",conductor_variables),"event","edat_pre","sexe")
formula<-formula_compare(x="regicor_vars_mis",y="edat_pre",taulavariables = conductor_variables) 
taula_REGICOR_controls_grups<-descrTable(formula,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F,show.all = T)

taula_REGICOR_controls_grups

```

```{r secundari3}

# Regicor per grups d'edad i sexe casos i controls

formula<-formula_compare(x="taula1",y="event",taulavariables = conductor_variables) 
dades_temp<-dades %>% 
  select(extreure.variables("taula1",conductor_variables),"event","sexe","edat_pre")

Taula3<-descrTable(event~.-sexe-edat_pre,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F)
Taula3.2<-strataTable(Taula3,"sexe")
Taula3.3<-strataTable(Taula3,"edat_pre")


```

>> Objetivo secundario 4

```{r secundario4}
formula<-formula_compare(x="regicor_vars",y="event",taulavariables = conductor_variables) 

# Població jove
dades_temp<-dades %>% filter(edat_grup=="Young:M<55/F<65") %>% 
  select(extreure.variables("regicor_vars",conductor_variables),"event","sexe","edat_pre")
taula4<-descrTable(event~.-sexe-edat_pre,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F)
taula4.1<-strataTable(taula4,"sexe")

# ORs
kable(M2.reg.vars.G1.ls$taula_OR,digits = 3) %>% kableExtra::kable_styling()

print(M2.reg.vars.G1.ls$auc)
M2.reg.vars.G1.ls$forest_plot
M2.reg.vars.G1.ls$ggplot_ROC
M2.reg.vars.G1.ls$auc_ci


# Población general
dades_temp<-dades  %>% 
  select(extreure.variables("regicor_vars",conductor_variables),"event","sexe","edat_pre")
taula4.3<-descrTable(event~.-sexe-edat_pre,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F)
taula4.4<-strataTable(taula4.3,"edat_pre")

# OR's 
kable(M2.reg.vars.G2.ls$taula_OR,digits = 3,caption="OR por componente de REGICOR población mayor") %>% kableExtra::kable_styling()
print(M2.reg.vars.G2.ls$auc)
M2.reg.vars.G2.ls$forest_plot
M2.reg.vars.G2.ls$ggplot_ROC
M2.reg.vars.G2.ls$auc_ci



```

>> Objetivo secundario 5

```{r secundario5}
formula<-formula_compare(x="regicor_plus",y="event",taulavariables = conductor_variables) 

# Població jove
dades_temp<-dades %>% filter(edat_grup=="Young:M<55/F<65") %>% 
  select(extreure.variables("regicor_plus",conductor_variables),"event","sexe","edat_pre","edat.cat5")
taula5<-descrTable(event~.-sexe-edat_pre-edat.cat5,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F)
taula5.1<-strataTable(taula5,"sexe")
taula5.1.1<-strataTable(taula5,"edat_pre")

# OR's 
kable(M3.reg.plus.G1.ls$taula_OR,digits = 3,caption="OR por componente de REGICOR población mayor") %>% kableExtra::kable_styling()
print(M3.reg.plus.G1.ls$auc)
M3.reg.plus.G1.ls$forest_plot
M3.reg.plus.G1.ls$ggplot_ROC
M3.reg.plus.G1.ls$auc_ci

# Población general
dades_temp<-dades  %>% 
  select(extreure.variables("regicor_vars",conductor_variables),"event","sexe","edat_pre","edat.cat5")
taula5.2<-descrTable(event~.-sexe-edat_pre-edat.cat5,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F)
taula5.3<-strataTable(taula5.2,"edat_pre")
taula5.4<-strataTable(taula5.2,"edat.cat5")

# OR's població gran
kable(M3.reg.plus.G2.ls$taula_OR,digits = 3,caption="OR por componente de REGICOR población mayor") %>% kableExtra::kable_styling()
print(M3.reg.plus.G2.ls$auc)
M3.reg.plus.G2.ls$forest_plot
M3.reg.plus.G2.ls$ggplot_ROC
M3.reg.plus.G2.ls$auc_ci


```

>> Objetivo secundario 6

```{r secundario6}
formula<-formula_compare(x="taula6",y="event",taulavariables = conductor_variables) 

# Població jove
dades_temp<-dades %>% filter(edat_grup=="Young:M<55/F<65") %>% 
  select(extreure.variables("taula6",conductor_variables),"event","sexe","edat_pre")
taula6<-descrTable(event~.-sexe-edat_pre,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F)
taula6.1<-strataTable(taula6,"sexe")
taula6.1.1<-strataTable(taula6,"edat_pre")

# Población general
dades_temp<-dades  %>% 
  select(extreure.variables("taula6",conductor_variables),"event","sexe","edat_pre")
taula6.2<-descrTable(event~.-sexe-edat_pre,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = T,show.ratio = F)
taula6.3<-strataTable(taula6.2,"edat_pre")



```


>> Tabla extra (tabla 9) impacto de factores de riesgo no incluidos en Regicor

```{r tablaextra1}

dades_temp<-dades %>% select(extreure.variables("taula9",conductor_variables),"event","sexe","edat_pre")

taula9<-descrTable(event~.-sexe-edat_pre,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = F,show.ratio = F)
taula9.1<-strataTable(taula9,"edat_pre")


# ORs (Model 4)
M4.reg.plus.G1.ls<-extreure_model_logistic(x="taula9",y="event",taulavariables=conductor_variables,dades=dades,valor_outcome=cat_event,conditional = T,strata = "caseid")

M4.reg.plus.G1.ls$taula_OR

kable(M4.reg.plus.G1.ls$taula_OR,digits = 2) %>% kableExtra::kable_styling()
print(M4.reg.plus.G1.ls$auc)
M4.reg.plus.G1.ls$forest_plot
M4.reg.plus.G1.ls$ggplot_ROC
M4.reg.plus.G1.ls$auc_ci


# Estratificado por edat i sexo edat_pre 

M4.reg.plus.grupos<-dades %>% split(.$edat_pre) %>%
  map(~ extreure_model_logistic(x="taula9",y="event",taulavariables=conductor_variables,dades=.x,valor_outcome=cat_event,conditional = T,strata = "caseid")[[1]]) 

M4.reg.plus.grupos$`Old man(>=55)`
M4.reg.plus.grupos$`Old woman(>=65)`
M4.reg.plus.grupos$`Young man(<55)`
M4.reg.plus.grupos$`Young woman(<65)`


```


```{r tablaextra2}

dades_temp<-dades %>% select(extreure.variables("taula9.2",conductor_variables),"event","sexe","edat_pre")
taula9.2<-descrTable(event~.-sexe-edat_pre,data=dades_temp,show.p.overall = F,hide.no = "No",show.n = F,show.ratio = F)
taula9.2.1<-strataTable(taula9.2,"edat_pre")

# ORs (Model 4)
# Generador de ORs (Model 4)
M4.2.reg.plus.G1.ls<-extreure_model_logistic(x="taula9.2",y="event",taulavariables=conductor_variables,dades=dades,valor_outcome=cat_event,conditional = T,strata = "caseid")


# Estratificado por edat i sexo edat_pre 
M4.2.reg.plus.grupos<-dades %>% split(.$edat_pre) %>%
  map(~ extreure_model_logistic(x="taula9",y="event",taulavariables=conductor_variables,dades=.x,valor_outcome=cat_event,conditional = T,strata = "caseid")[[1]]) 

# Estratificado por edat i sexo edat_pre 

```



>> Tabla extra 2 (tabla 10) Razón de odds (OR) de ECV según REGICOR 

```{r tablaextra3}

# Estimación de OR de evento en función de Regicor en toda la población 
M5.reg.cont<-extreure_model_logistic(x="regicor_alone_continu",y="event",taulavariables=conductor_variables,dades=dades,valor_outcome=cat_event,conditional = T,strata = "caseid")
M5.reg.cat<-extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=dades,valor_outcome=cat_event,conditional = T,strata = "caseid")


# Estratificado por sexo (H/M)
M5.reg.cont.sex<-dades %>% split(.$sexe) %>%
  map(~ extreure_model_logistic(x="regicor_alone_continu",y="event",taulavariables=conductor_variables,dades=.x,valor_outcome=cat_event,conditional = T,strata = "caseid")[[1]]) 

M5.reg.cat.sex<-dades %>% split(.$sexe) %>%
  map(~ extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=.x,valor_outcome=cat_event,conditional = T,strata = "caseid")[[1]]) 


# Estratificado por edat_pre (H/M)
M5.reg.cont.edat_pre<-dades %>% split(.$edat_pre) %>%
  map(~ extreure_model_logistic(x="regicor_alone_continu",y="event",taulavariables=conductor_variables,dades=.x,valor_outcome=cat_event,conditional = T,strata = "caseid")[[1]]) 

M5.reg.cat.edat_pre<-dades %>% split(.$edat_pre) %>%
  map(~ extreure_model_logistic(x="regicor_alone",y="event",taulavariables=conductor_variables,dades=.x,valor_outcome=cat_event,conditional = T,strata = "caseid")[[1]]) 


```


# 7. Salvar objectes

```{r salvar_objectes}


if (params$mostra) dir_Rdata<-"output/R_data/test" else dir_Rdata<-"output/R_data"

# Netejo ambient
gc()
rm(list=c("dadesevents","dades_temp","flow_chart_antic","flow_chart_antic2","temp_dt","params"))
rm(params)

save.image(file=here::here(dir_Rdata,"output_PRECAV.Rdata"))
# saveRDS(dades,here::here(dir_Rdata,"dades.Rds"))



```
